\name{bmds-package}
\alias{bmds-package}
\alias{bmds}
\docType{package}
\title{
\packageTitle{bmds}
}
\description{
\packageDescription{bmds}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bmds}
\packageIndices{bmds}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bmds}

Maintainer: \packageMaintainer{bmds}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
